requirements.txt

# Tkinter 
tkinter
customtkinter


pandas

# Plotting/graphs
matplotlib

# Environment variables
python-dotenv

# Yahoo imports
yfinance